requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
django==4.1
django-environ==0.10.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
pytz==2021.3
gunicorn==20.0.4
drf-spectacular>=0.22.1,<0.23
djangorestframework>=3.13.1,<3.14
pillow==5.0.0
django-cors-headers==3.14.0
stripe==2.63.0
